Convert token_type_ids to Long

+++ spacy_transformers/layers/transformer_model.py
@@ -210,7 +210,7 @@ def _convert_transformer_inputs(model, wps: WordpieceBatch, is_train):
         "attention_mask": xp2torch(wps.attention_mask).to(device=hf_device),
     }
     if wps.token_type_ids is not None:
-        kwargs["token_type_ids"] = xp2torch(wps.token_type_ids).to(device=hf_device)
+        kwargs["token_type_ids"] = xp2torch(wps.token_type_ids).long().to(device=hf_device)
     return ArgsKwargs(args=(), kwargs=kwargs), lambda dX: []
